CI-246: Moved directory sys_svg to the images and deleted non-used svg file
